refactor: Extract timed decorator
